6_LAB/3_ZAD/backend/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
requests==2.26.0
